AI/vpoc/agents/validation_agent.py
import logging
import typing
from google.adk.agents.base_agent import BaseAgent
from google.adk.agents.invocation_context import InvocationContext
from google.adk.events.event import Event
from google.genai import types

from .base import VPOCMixin
from core.utils import PromptLoader

logger = logging.getLogger(__name__)

class ValidationAgent(VPOCMixin, BaseAgent):
    """
    Validation Agent.
    Executes PoCs in a sandboxed environment and analyzes results.
    """
    description: str = "Executes PoCs in a hardened sandbox and confirms security impact."

    def __init__(self, **data: typing.Any):
        super().__init__(**data)
        self.prompt_loader = PromptLoader()

    async def _run_async_impl(
        self, ctx: InvocationContext
    ) -> typing.AsyncGenerator[Event, None]:
        yield Event(
            author=self.name,
            content=types.Content(parts=[types.Part(text="Validation Agent executing PoC...")]),
        )
        # TODO: Implement SandboxRunner integration and LLM-driven outcome analysis.
        
        yield Event(
            author=self.name,
            content=types.Content(parts=[types.Part(text="Validation Agent: Execution complete.")]),
        )
