AI/vpoc/pyproject.toml
[project]
name = "vpoc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "docker>=7.1.0",
    "fastapi>=0.135.3",
    "google-adk>=1.29.0",
    "httpx>=0.28.1",
    "litellm>=1.83.4",
    "python-multipart>=0.0.26",
    "sqlmodel>=0.0.38",
    "textual>=8.2.3",
    "tomli>=2.4.1",
    "uvicorn[standard]>=0.44.0",
]

[dependency-groups]
dev = [
    "black>=26.3.1",
    "flake8>=7.3.0",
    "mypy>=1.20.0",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "respx>=0.23.1",
]
